chore(team): split mailbox remote tests
